apps/comments/requirements.txt
fastapi[standard]
tortoise-orm[asyncmy]
aerich[toml]
aiofiles
httpx[http2]
cryptography